try again

+++ libs/fmts/fmts/__init__.py
@@ -1298,7 +1298,7 @@ def dedent(string: str) -> str:
 
 
 def timestamp(
-    ts: Optional[Union[float, datetime]] = None, utc_offset: Optional[int] = None
+    ts: Optional[Union[float, datetime]] = None, utc_offset: Optional[float] = None
 ) -> str:
     """Time stamp string w/ format yyyymmdd-HHMMSS
     Defaults
@@ -1332,7 +1332,7 @@ def timestamp(
 
 
 def long_timestamp_string(
-    timestamp_sec: float, utc_offset: Optional[int] = None
+    timestamp_sec: float, utc_offset: Optional[float] = None
 ) -> str:
     """Return a 'long-form' timestamp string given epoch-seconds float and utc offset hours as int"""
     if utc_offset == None:
